[UPDATE] add utils folder and move decodeJWT file to utils folder
